Safari-609.1.11.1/LayoutTests/fast/history/history-subframe-with-name-expected.txt

main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
Verifies that we can go back and forward from the top level of a frame set where a window changes its name after the initial load.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 2 is currentPageId
PASS 3 is currentPageId
PASS 2 is currentPageId
PASS 3 is currentPageId
PASS Complete: navigated through all the states
